infra/labelstudio/requirements.txt
label-studio-sdk>=0.0.32,<0.1
python_dotenv>=1.0.1,<1.1
